+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Next up
+
+- Foe Slayer effect from @motomoto0295
+
 # 5.2.24
 
 - Darkness Spell: in v12 will now use darkness lightsource. A Simple Macro will be added to the description to toggle placement/removal of the darkness light if not using midi-qol.